requirements.txt
fastapi
uvicorn
python-keycloak
requests
httpx
pydantic
python-dotenv
alembic
sqlalchemy
asyncpg
